fix: prevent deleting the Email field in the fields modal
